Add --merge-split and --no-prune options.
The --merge-split=n allows movies that are split for news to be merged.
The n is maximum duration in minutes that the news can be. If two
programmes with same programid are separated by a programme of
this duration or less then they will be assumed to be part of the same
movie and so joined together as one programme entry.

The default value is 0 (no merging of programmes).

The --no-prune is mainly used when debugging along with --no-download.