Add optional caching of xmltv programme data to redis.
The re-generation of files is slow on my machine. So
if redis is installed, we allow the user to add
"--cache-driver=Redis" (with a capital "R") and
it will cache the programme data. Then, on subsequent
runs, this data is reused, significantly reducing
overhead for subsequent runs since we download
+n days, hence the data for those programmes
does not need to be regenerated for n days.

Requires (optional) Perl CHI module, Perl CHI Redis
driver, and a redis server running locally.