Include a cache for program details.
Many programmes are repeated during a week. So we cache the
program details to speed up those lookups. However, we keep the
program (schedule) cache since this avoids having to decode JSON
to do the processing.

We use an L1 cache for the programme details cache to avoid having
to query the external cache.