Delete channels not in lineup during fetch.
Seems like a server-side bug that for very small subset
of stations, SD says a station is in our lineup, then a
fetch says the station is not in any of our lineups.

So we now delete the station from our DB. However a second
download then repopulates the lineup with the station.