Schedules now start from "now" rather than start of day.
Previously we'd round to nearest day (ignore time). But that means
if you run the grabber in the evening, you get the entire of
today's listings (including morning and afternoon) which is not
useful.

So, we start from "now", but we subtract a few hours since
otherwise if a programme is half way through then its starttime
would be in the past so the programme would not be included.

Of course, the opposite problem is that now if you ran at
midnight, you now get a few programmes from yesterday...