Add --channel-short-days-exclude-regex and --channel-short-days.
This allows you to download a longer guide for "good" channels
and a shorter programme guide for channels that are not interesting
to you. This helps save memory on PVRs and clients.

For example:
--channel-short-days=3 --channel-short-days-exclude-regex="PBS|Shopping"
Will give you a short programme guide of three days for all channels
except for channels that match the regex "PBS" or "Shopping".