Add script for supplying Tvheadend 4.3+ with extra fanart for Kodi 18+.
Tvheadend 4.3+ and Kodi 18 support fanart on recordings.  Since xmltv
does not provide a good portable mechanism for passing fanart, we
provide a new script to set it inside Tvheadend.

This is called tv_meta_az_sd.py and should be placed in the same
directory as Tvheadend's "tvhmeta". The Tvheadend's script
then uses our script to supply extra fanart to Kodi.

This can be configured via the tvheadend pre-processor command (in
Configuration->Recording->pre_processor command) with the arguments
"/usr/local/bin/tvhmeta --uuid %U" (without the quotes, and fixing
path appropriately).

For example:
/usr/local/bin/tvhmeta --uuid %U

The extra fanart functionality in Tvheadend is currently undergoing a
variety of changes so future versions of Tvheadend are likely to
automatically set fanart without needing a pre-processor command,
or may break this new feature, so check the NEWS frequently.