Initial support for IMDB.
This allows us to have ratings for shows/episodes
and not just for movies. We also now pass the IMDB
number through in the metadata to the PVR.

Add option --use-imdb and --prefer-imdb-rating.
Star ratings are now out of 5.

We use mysql for the imdb database since SQLite
was too slow for regular importing of the bulk data.