Add new option '--update-category-with-grabber'.
This adds an "xmltv" tag to the list of categories for a programme.
The user can then schedule a recording in the PVR that has to
have this category (along with other criteria such as title) and
so avoid recording programmes where OTA broadcast has provided
an EIT.

For example, in my region we have a number of channels for which
SchedulesDirect provides no information, therefore you need a mixed
system of SchedulesDirect and OTA EIT.

But if you have channel "XYZ" and a retune might provide a new
timeshift variant "XYZ+1". However, SchedulesDirect might not
have EPG information for the channel. Thus the PVR might populate
(incomplete) OTA information for the "XYZ+1" channel. So if you have a
rule "record title=Fred" then it might also record programmes on the
timeshift OTA EIT channel.

By updating the recording rule to "record title=Fred and category=xmltv"
we avoid any possibility of recording programmes that aren't
provided by SchedulesDirect, whilst still allowing the user to see
OTA EIT information for channels not provided by SchedulesDirect.

Also tidied sorting of category information.